requirements.txt
Flask==2.0.1
Flask-Cors==4.0.0
markdown==3.3.7
python-dotenv==0.19.0
requests==2.28.1
PyMuPDF==1.20.2  # For PDF text extraction